move over sprite tests
