+++ sprites/lib/compass/sprites/version.rb
@@ -1,0 +1,5 @@
+module Compass
+  module Sprites
+    VERSION = File.read(File.join(File.dirname(__FILE__), "..", "..", "..", "VERSION")).strip
+  end
+end
