+++ sprites/test/integrations/sprites_test.rb
