+++ sprites/test/units/image_row_test.rb
