+++ sprites/test/units/image_test.rb
