+++ sprites/test/units/images_test.rb
