+++ sprites/test/units/importer_test.rb
