+++ sprites/test/units/row_fitter_test.rb
