+++ sprites/test/units/sprite_command_test.rb
