+++ sprites/test/units/sprite_map_test.rb
